requirements.txt
tensorboard==2.1.0
PyYAML==5.1
tqdm==4.43.0
numpy==1.18.0

